requirements.txt
# DO NOT include azure-functions-worker in this file
# The Python Worker is managed by Azure Functions platform
# Manually managing azure-functions-worker may cause unexpected issues

azure-functions
azure-common==1.1.28
azure-core==1.24.2
azure-cosmos==4.3.0
azure-identity==1.10.0
azure-keyvault==4.2.0
azure-keyvault-certificates==4.4.0
azure-keyvault-keys==4.5.1
azure-keyvault-secrets==4.4.0
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.1.0
click==8.1.3
colorama==0.4.5
cryptography==37.0.4
Flask==2.1.3
Flask-Cors==3.0.10
idna==3.3
importlib-metadata==4.12.0
isodate==0.6.1
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.1.0
MarkupSafe==2.1.1
mongoengine==0.24.2
msal==1.18.0
msal-extensions==1.0.0
msrest==0.7.1
numpy==1.23.1
oauthlib==3.2.0
pandas==1.4.3
portalocker==2.5.1
pycparser==2.21
PyJWT==2.4.0
pymongo==4.2.0
python-dateutil==2.8.2
pytz==2022.1
requests==2.28.1
requests-oauthlib==1.3.1
six==1.16.0
threadpoolctl==3.1.0
typing-extensions==4.3.0
urllib3==1.26.11
Werkzeug==2.1.2
zipp==3.8.1
